+++ bin/csm.js
@@ -1,4 +1,5 @@
 #! /usr/bin/env node
+"use strict";
 
 import checkStatsModules from "../lib/check-stats-modules.js";
