+++ lib/json2table.js
@@ -1,26 +1,29 @@
 "use strict";
 
-var JSONTable = require('json-table');
-module.exports = function (stats) {
-  return new JSONTable(stats, {
-    chars: {
-      'top': '═',
-      'top-mid': '╤',
-      'top-left': '╔',
-      'top-right': '╗',
-      'bottom': '═',
-      'bottom-mid': '╧',
-      'bottom-left': '╚',
-      'bottom-right': '╝',
-      'left': '║',
-      'left-mid': '╟',
-      'mid': '─',
-      'mid-mid': '┼',
-      'right': '║',
-      'right-mid': '╢',
-      'middle': '│'
-    }
-  }, function (table) {
-    return table;
-  });
+import JSONTable from "json-table";
+
+export default (stats) => {
+  return new JSONTable(
+    stats,
+    {
+      chars: {
+        top: "═",
+        "top-mid": "╤",
+        "top-left": "╔",
+        "top-right": "╗",
+        bottom: "═",
+        "bottom-mid": "╧",
+        "bottom-left": "╚",
+        "bottom-right": "╝",
+        left: "║",
+        "left-mid": "╟",
+        mid: "─",
+        "mid-mid": "┼",
+        right: "║",
+        "right-mid": "╢",
+        middle: "│",
+      },
+    },
+    (table) => table
+  );
 };
