+++ lib/utils.js
@@ -1,15 +1,17 @@
 "use strict";
 
-var args = require("./args");
-var _require = require("date-fns"),
-  format = _require.format,
-  parseISO = _require.parseISO,
-  isValid = _require.isValid,
-  addDays = _require.addDays,
-  addWeeks = _require.addWeeks,
-  addMonths = _require.addMonths,
-  addYears = _require.addYears,
-  getYear = _require.getYear;
+import chalk from "chalk";
+import yargs from "yargs";
+import {
+  format,
+  parseISO,
+  isValid,
+  addDays,
+  addWeeks,
+  addMonths,
+  addYears,
+  getYear,
+} from "date-fns";
 
 /**
  * show help
@@ -17,18 +19,18 @@ var _require = require("date-fns"),
  * @param  {String} help message if specifying a message
  * @return {String} full help message
  */
-exports.show_help = function (text) {
-  if (text) console.log(require("chalk").yellow.bold(text));
-  require("yargs").showHelp();
+const showHelp = (text) => {
+  if (text) console.log(chalk.yellow.bold(text));
+  yargs(process.argv.slice(2)).showHelp();
 };
 
 /**
  * get start date by option
  *
- * @param  {String} option
- * @return {String} start_date
+ * @param  {String} args
+ * @return {String} start date 'yyyy-MM-dd' format
  */
-exports.get_start_date = function () {
+const getStartDate = (args) => {
   // check module names
   if (args._.length === 0) {
     return "Please enter the module names at least one. \n";
@@ -43,13 +45,13 @@ exports.get_start_date = function () {
     }
   }
   if (args.m) {
-    return format(addMonths(new Date(), -1), "yyyy-MM-dd");
+    return format(addMonths(new Date(), 0), "yyyy-MM-dd");
   }
   if (args.y) {
     return format(addYears(new Date(), -1), "yyyy-MM-dd");
   }
   if (args.t) {
-    return "".concat(getYear(new Date()), "-01-01");
+    return `${getYear(new Date())}-01-01`;
   }
   if (args.w) {
     return format(addWeeks(new Date(), -1), "yyyy-MM-dd");
@@ -60,10 +62,10 @@ exports.get_start_date = function () {
 /**
  * get end date by option
  *
- * @param  {String} option
- * @return {String} end_date
+ * @param  {String} args
+ * @return {String} end date 'yyyy-MM-dd' format
  */
-exports.get_end_date = function () {
+const getEndDate = (args) => {
   // check module names
   if (args._.length === 0) {
     return "Please enter the module names at least one. \n";
@@ -93,12 +95,15 @@ exports.get_end_date = function () {
  * @param  {Object} second stats object in processing sort function
  * @return {Number} 0: equal, 1: a < b, -1: a > b
  */
-exports.compare = function (a, b) {
+const compare = (a, b) => {
   if (a.downloads < b.downloads) {
     return 1;
   }
   if (a.downloads > b.downloads) {
     return -1;
   }
+
   return 0;
 };
+
+export { showHelp, getStartDate, getEndDate, compare };
