--- src/lib/json2table.js
@@ -1,27 +1,0 @@
-import JSONTable from "json-table";
-
-export default (stats) => {
-  return new JSONTable(
-    stats,
-    {
-      chars: {
-        top: "═",
-        "top-mid": "╤",
-        "top-left": "╔",
-        "top-right": "╗",
-        bottom: "═",
-        "bottom-mid": "╧",
-        "bottom-left": "╚",
-        "bottom-right": "╝",
-        left: "║",
-        "left-mid": "╟",
-        mid: "─",
-        "mid-mid": "┼",
-        right: "║",
-        "right-mid": "╢",
-        middle: "│",
-      },
-    },
-    (table) => table
-  );
-};
